fix move hdi storage account code issue
